gym website

+++ renuka_gym/src/main/java/com/xworkz/gym/dto/AdminRegistrationDTO.java
@@ -25,7 +25,7 @@ public class AdminRegistrationDTO {
     private String gender;
     private int weight;
     private int age;
-    private double height;
+    private float height;
 
 
 }
